增加Base64编码；CRefStr增加转换到NT字符串结构方法，初始化时改为设置结尾NULL；修正phnt中RTL_BUFFER定义；FLAC标签图片写入选项；增加枚举进程，取进程路径，NT路径转DOS路径

+++ eck/CVectorDrawPanel.h
@@ -539,7 +539,7 @@ public:
 		if (cchText < 0)
 			cchText = (int)wcslen(pszText);
 		if (FAILED(hr = g_pDwFactory->CreateTextLayout(pszText, (UINT32)cchText,
-			m_pTf, m_cxClient, m_cyClient, &pLayout)))
+			m_pTf, (float)m_cxClient, (float)m_cyClient, &pLayout)))
 			return hr;
 		ID2D1PathGeometry1* pPath;
 		if (FAILED(hr = GetTextLayoutPathGeometry(pLayout.Get(), m_pDC1, 0.f, 0.f, pPath)))
